Snythetic-data-generation/Create_Seg_Dataset/requirements.txt
# cmake
pyarrow==17.0.0
torch
torchvision
transformers
diffusers
accelerate
datasets
Pillow
huggingface_hub
gguf
tqdm
protobuf
sentencepiece 
safetensors